- fix configuration for bytearray to wav conversion & normalize wav data - adjust sample rate, audio format, buffer size -adjust time tracker

+++ app/src/main/java/com/apicta/myoscopealert/utils/ArrayReceiver.kt
@@ -3,9 +3,9 @@ package com.apicta.myoscopealert.utils
 object ArrayReceiver {
 
     //for plot
-    val pcgArray = arrayListOf<Float>()
+    val pcgArray = arrayListOf<Int>()
     val timeArray = arrayListOf<Float>()
 
-    var pcgArrayNormalize = arrayListOf<Float>()
+    var pcgArrayNormalize = arrayListOf<Int>()
 
 }
